Refactor EditAlbum and NewAlbum to use a shared form
